Fixes error when you set email on a phone login
Currently, if a server has phone login enabled, you can register with
your phone, and if on the profile page you try to set your email, you
get this error:

```
  File ".../docassemble/webapp/users/forms.py", line 197, in validate
    existing_user = db.session.execute(select(UserModel).filter_by(email=self.email.data, active=True)).scalar()
NameError: name 'db' is not defined
```

This happens because the `db` module isn't actually loaded in this
scope. Loading the db_object fixes that.